requirements.txt
Flask==0.10.1
Flask-Mail==0.9.1
Flask-Script==2.0.5
Jinja2==2.7.3
MarkupSafe==0.23
PyYAML==3.11
SQLAlchemy==0.9.7
Werkzeug==0.9.6
blinker==1.3
coverage==3.7.1
flake8==2.2.3
gnureadline==6.3.3
ipython==2.2.0
itsdangerous==0.24
jedi==0.8.1-final0
mccabe==0.2.1
mysqlclient==1.3.3
nose==1.3.4
pep8==1.5.7
pyflakes==0.8.1
requests==2.4.1
beautifulsoup4==4.3.2
bottlenose==0.6.2